+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 7.1.0
+
+## Features
+
+- Can use the `_` alias to refer to the
+  [current process's Node.js version](/README.md#supported-aliases)
+- Can use the `.` alias to refer to the
+  [current project's Node.js version](/README.md#supported-aliases) using its
+  `.nvmrc`, `.node-version` or `.naverc`. The current directory can be changed
+  using the [`cwd` option](/README.md#cwd).
+
 # 7.0.1
 
 ## Bug fixes